refactor(client): P5b — handler bodies to native_tools domain modules (4 of 5 domains) (#137)
* refactor(client): P5b(1/5) — scheduling handlers to native_tools/scheduling.py

Verbatim body moves: schedule_task/list/update/delete_schedule/parse_time
+ _validate_schedule_payload/_extract_tool_input_from_steps into
SchedulingTools. Bot keeps one-line delegates (dispatch host + test seam;
smoke tests call _validate_schedule_payload directly).

* refactor(client): P5b(2/5) — knowledge/history/audit-search handlers to native_tools/knowledge.py

Verbatim moves; _knowledge_store is provider-backed on the domain class
(the bot's knowledge setter replaces the store at runtime).

* refactor(client): P5b(3/5) — channel-ops handlers to native_tools/channel_ops.py

Verbatim moves: purge, read_channel, add_reaction, create_poll,
set_permission. get_channel passed as the bot's bound method.

* refactor(client): P5b(4/5) — media/file handlers to native_tools/media.py

Verbatim moves: browser_screenshot, generate_file, post_file,
analyze_image (vision marker dict preserved), generate_image,
_detect_image_type (bot keeps a staticmethod alias). get_config
provider-backed; asyncio.TimeoutError normalized to TimeoutError in the
moved post_file body (builtin alias, new-file lint).

* refactor(client): P5b(5/5) — RFC R4 amendment + import cleanup

Agents/tasks/loops handler domain deferred to a P5c rider after P8 (RFC
R4): those handlers close over the loop pipeline (_run_loop_iteration,
llm_client, lifecycle plumbing) that P7/P8 restructure — moving them
first would mean moving them twice across the campaign's riskiest phases.
Dispatch is unaffected (late-bound host). base64 import removed (users
moved to media.py).
